Preserve folder/file structure as before inside the "src" folder
