--- packages/js/e2e-core-tests/index.js
@@ -1,6 +1,0 @@
-/*
- * Internal dependencies
- */
-const allSpecs = require( './src' );
-
-module.exports = allSpecs;
