+++ packages/js/e2e-core-tests/src/specs/activate-and-setup/activate.test.js
