+++ packages/js/e2e-core-tests/src/specs/activate-and-setup/setup.test.js
