+++ packages/js/e2e-core-tests/src/specs/api/coupon.test.js
