+++ packages/js/e2e-core-tests/src/specs/api/external-product.test.js
