+++ packages/js/e2e-core-tests/src/specs/api/telemetry.test.js
