+++ packages/js/e2e-core-tests/src/specs/api/variable-product.test.js
