+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-analytics-page-loads.test.js
