+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-coupon-new.test.js
