+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-extensions-connect-wccom.test.js
