+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-order-apply-coupon.test.js
