+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-product-edit-details.test.js
