+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-product-new.test.js
