+++ packages/js/e2e-core-tests/src/specs/merchant/wp-admin-settings-shipping-classes.test.js
