+++ packages/js/e2e-core-tests/src/specs/shopper/front-end-checkout-coupons.test.js
