+++ packages/js/e2e-core-tests/src/specs/shopper/front-end-checkout-create-account.test.js
