+++ packages/js/e2e-core-tests/src/specs/shopper/front-end-order-email-receiving.test.js
