+++ packages/js/e2e-core-tests/src/specs/shopper/front-end-single-product.test.js
