+++ packages/js/e2e-core-tests/src/specs/shopper/front-end-variable-product-updates.test.js
